+++ HISTORY.md
@@ -1,5 +1,8 @@
 ### mlpack ?.?.?
 ###### ????-??-??
+  * Added Extra Trees Algorithm (#2883). Currently, it can be used using the
+    class `mlpack::tree::ExtraTrees`, but only through C++.
+
   * Add Flatten T Swish activation function (`flatten-t-swish.hpp`)
 
   * Added warm start feature to Random Forest (#2881); this feature is